+++ R/new_project.R
@@ -1,36 +1,30 @@
 #'creates folder for a new project and adds templates for all the necessary files.
 #'@param project_id character. ID of the project, e.g. 2021-03-COR. see description for details.
-#'@param data_folder character. path to data folder starting at root of the project. defaults to here::here()
+#'@param name character. short name of the project. defaults to "".
+#'@param data_folder character. path to data folder starting at root of the project. defaults to "."
 #'@details the project id is composed by the year-month of when the project was started (~when the kickoff took place) and a three-letter, uppercase
 #' abbreviation for the organization. If there are two projects with the same organization starting in the same month, 
 #' two letters of the abbreviation should be used for the organization, and one for the content of the project. e.g. 2021-03-COV for a visualization project
 #' with CorrelAid and 2021-03-COA for an automation project with CorrelAid.
 #'@export  
-new_project <- function(project_id, data_folder = here::here()) {
+
+new_project <- function(project_id, name, data_folder = here::here()) {
   assert_project_id(project_id)
   usethis::ui_info(glue::glue("processing {project_id}"))
-  
+
   dir.create(fs::path(data_folder, project_id), showWarnings = FALSE)
   
-  # meta data file (meta.json)
-  template_meta <- get_meta_template()
-  
-  # populate with data 
-  template_meta$project_id <- project_id
-  template_meta$year <- stringr::str_sub(project_id, 1, 4)
-  template_meta$start <- stringr::str_sub(project_id, 1, 7)
-  
-  meta_path <- fs::path(data_folder, project_id, "meta.json")
+  proj <- Project$new(project_id, name)
+  proj_obj_path <- fs::path(data_folder, project_id, glue::glue("{project_id}.rds"))
 
   answer <- TRUE
-  if (file.exists(meta_path)) {
-    usethis::ui_warn("meta.json already exists in {meta_path}")
+  if (file.exists(proj_obj_path)) {
+    usethis::ui_warn(glue::glue("{project_id}.rds already exists in {proj_obj_path}"))
     answer <- usethis::ui_yeah("Do you want to overwrite it?", yes = "Yes", no = "No", shuffle = FALSE)
-    
   }
   if (answer) {
-    readr::write_lines(jsonlite::prettify(jsonlite::toJSON(template_meta, auto_unbox = TRUE)), meta_path)
-    usethis::ui_done(glue::glue("created meta.json at {meta_path}."))
+    readr::write_rds(proj, proj_obj_path)
+    usethis::ui_done(glue::glue("created {project_id}.rds at {proj_obj_path}."))
   }
   
   # markdown files
@@ -51,30 +45,79 @@ new_project <- function(project_id, data_folder = here::here()) {
       }
     )
   }
+  use_team_selection_workflow(project_id, data_folder)
+  use_r6_templates(project_id, data_folder)
 }
 
-
-#'Use fill project script template
+#'Use R6 templates
 #'@param project_id project id in path form, e.g. 2020-11-COR
-#'@param data_folder character. path to data folder starting at root of the project. defaults to "", i.e. root
+#'@param data_folder character. path to data folder starting at root of the project. defaults to ".", i.e. root
 #'@export 
-use_fill_project <- function(project_id, data_folder = "") {
-  
+use_r6_templates <- function(project_id, data_folder = ".") {
   usethis::use_template(
-    "fill_project.R",
-    save_as = fs::path(data_folder, project_id, "fill_project.R"),
+    "project_data_public.R",
+    save_as = fs::path(data_folder, project_id, glue::glue("{project_id}_project_data_public.R")),
     data = list(project_id = project_id),
     package = "projectutils",
     open = TRUE
   )
+  usethis::use_template(
+    "project_data_sensitive.R",
+    save_as = fs::path(data_folder, project_id, glue::glue("{project_id}_project_data_sensitive.R")),
+    data = list(project_id = project_id),
+    package = "projectutils",
+    open = FALSE
+  )
 }
 
-get_meta_template <- function() {
-  path <- tryCatch(
-    fs::path_package(package = "projectutils", "templates", "template_meta.json"), # installed in library
-    error = function(e) fs::path_package(package = "projectutils", "inst", "templates", "template_meta.json") # development mode
-  )
+
+#'Use team selection workflow
+#'@param project_id project id in path form, e.g. 2020-11-COR
+#'@param data_folder character. path to data folder starting at root of the project. defaults to ".", i.e. root
+#'@export 
+use_team_selection_workflow <- function(project_id, data_folder = ".") {
   
-  j <- jsonlite::read_json(path)
-  return(j)
+  project_folder <- fs::path(data_folder, project_id)
+  
+  # create subfolder for team selection if not already there
+
+  team_selection_folder <- fs::path(data_folder, project_id, "team_selection")
+  if (!dir.exists(team_selection_folder)) {
+    dir.create(team_selection_folder)
+    dir.create(fs::path(team_selection_folder, "data"))
+  }
+  
+  # use templates
+  # we add the three-letter prefix 
+  # to make file names distinguishable in RStudio when more then one project is managed at the same time
+  prefix <- tolower(stringr::str_extract(project_id, '[:alpha:]+')) 
+  usethis::use_template(
+    "send_confirmation_emails.R",
+    save_as = fs::path(team_selection_folder, glue::glue("02_{project_id}_send_confirmation_emails.R")),
+    data = list(project_id = project_id),
+    package = "projectutils",
+    open = FALSE
+  )
+  usethis::use_template(
+    "template_application_single.Rmd",
+    save_as = fs::path(team_selection_folder, glue::glue("zzz_template_application_single.Rmd")),
+    package = "projectutils",
+    open = FALSE
+  )
+
+  usethis::use_template(
+    "template_applications_report.Rmd",
+    save_as = fs::path(team_selection_folder, glue::glue("{project_id}_applications_report.Rmd")),
+    data = list(project_id = project_id),
+    package = "projectutils",
+    open = FALSE
+  )
+
+  usethis::use_template(
+    "prepare_team_selection.R",
+    save_as = fs::path(team_selection_folder, glue::glue("01_{project_id}_prepare_team_selection.R")),
+    data = list(project_id = project_id),
+    package = "projectutils",
+    open = TRUE
+  )
 }